--- src/main/kotlin/lesson_11/User.kt
@@ -1,12 +1,0 @@
-package lesson_11
-
-class User(
-    val uid: String,
-    val login: String,
-    val password: String,
-    val email: String,
-) {
-    override fun toString(): String {
-        return "User(uid='$uid', login='$login', login='$password', email='$email')"
-    }
-}
