+++ src/main/kotlin/lesson_11/lesson11_task1.kt
@@ -1,4 +1,13 @@
-package lesson_11
+class User(
+    val uid: String,
+    val login: String,
+    val password: String,
+    val email: String,
+) {
+    override fun toString(): String {
+        return "User(uid='$uid', login='$login', login='$password', email='$email')"
+    }
+}
 
 fun main() {
     val user1 = User(
